DEV-81 Moved Transmit Table
