+++ src/components/FormComponents/Button.module.scss
@@ -6,7 +6,7 @@
 
   text-align: center;
   @include styles.default-text-2;
-  background-color: styles.$custom-orange;
+  background-color: styles.$orange;
   border-radius: styles.$default-border-radius;
   color: white;
