+++ src/components/TransmitTable/OrderDropdown/OrderDropdown.module.scss
@@ -3,7 +3,7 @@
 #wrapper {
   flex: 0 0 auto;
   width: 100%;
-  border: 1px solid styles.$custom-orange;
+  border: 1px solid styles.$orange;
   border-radius: styles.$default-border-radius;
   overflow: hidden;
 }
