+++ src/components/TransmitTable/OrderFields/OrderFields.module.scss
@@ -4,5 +4,5 @@
   flex: 1 1 0;
   width: 100%;
   @include styles.default-text-2;
-  border-top: 1px solid styles.$custom-orange;
+  border-top: 1px solid styles.$orange;
 }
