+++ src/components/TransmitTable/OrderHeader/OrderHeader.module.scss
@@ -14,26 +14,13 @@ $header-color: #ffe4cc;
   background-color: $header-color;
   @include styles.default-text-2;
 
-  #caret {
-    grid-area: caret;
-    justify-self: center;
-    width: 1rem;
-    height: 1rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex: 0 0 auto;
-    color: styles.$custom-orange;
-
-    //TODO: arreglar caret, que ocupe el 100%
-    #caret > * {
-      width: 100%;
-      height: 100%;
-    }
+  .caretWrapper > * {
+    color: styles.$orange;
   }
+
   #name {
     grid-area: name;
-    color: styles.$custom-orange;
+    color: styles.$orange;
     font-weight: 500;
   }
 }
